Refactoring: Merge SchemaHelper into DBSQLOperation
